# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TMCpos Textile POS/ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List all purchases
      parameters:
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create a purchase transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Created purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Get a purchase with line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseDetail"

  # Mutations (Mutasi)
...
  /receivables:
    get:
      operationId: listReceivables
      tags: [receivables]
      summary: List all receivables
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: overdue
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of receivables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receivable"
  /receivables/{id}:
    get:
      operationId: getReceivable
      tags: [receivables]
      summary: Get receivable with payment history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receivable detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceivableDetail"
  /receivables/{id}/payments:
    post:
      operationId: addReceivablePayment
      tags: [receivables]
      summary: Add a payment to a receivable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # Payables (Hutang)
...
components:
  schemas:
...
    Purchase:
      type: object
      required: [id, invoiceNumber, supplierId, paymentType, totalAmount, status, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: ["string", "null"]
        paymentType:
          type: string
        totalAmount:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PurchaseItem:
      type: object
      required: [id, purchaseId, productId, rolls, meters, pricePerMeter, subtotal]
      properties:
        id:
          type: integer
        purchaseId:
          type: integer
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        rollId:
          type: ["integer", "null"]
        rolls:
          type: number
        meters:
          type: number
        pricePerMeter:
          type: number
        subtotal:
          type: number
    PurchaseItemInput:
      type: object
      required: [productId, rolls, meters, pricePerMeter, subtotal]
      properties:
        productId:
          type: integer
        rollId:
          type: integer
        rolls:
          type: number
        meters:
          type: number
        pricePerMeter:
          type: number
        subtotal:
          type: number
        barcode:
          type: string
    PurchaseDetail:
      type: object
      required: [id, invoiceNumber, supplierId, paymentType, totalAmount, status, createdAt, items]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: ["string", "null"]
        paymentType:
          type: string
        totalAmount:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
    PurchaseInput:
      type: object
      required: [supplierId, paymentType, items]
      properties:
        supplierId:
          type: integer
        paymentType:
          type: string
        dueDate:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItemInput"
...
    Receivable:
      type: object
      required: [id, saleId, customerId, totalAmount, paidAmount, remainingAmount, status, createdAt]
      properties:
        id:
          type: integer
        saleId:
          type: integer
        invoiceNumber:
          type: ["string", "null"]
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        totalAmount:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        isOverdue:
          type: boolean
        createdAt:
          type: string
    ReceivableDetail:
      type: object
      required: [id, saleId, customerId, totalAmount, paidAmount, remainingAmount, status, createdAt, payments]
      properties:
        id:
          type: integer
        saleId:
          type: integer
        invoiceNumber:
          type: ["string", "null"]
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        totalAmount:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        isOverdue:
          type: boolean
        createdAt:
          type: string
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
...
    Payment:
      type: object
      required: [id, amount, paymentMethod, paidAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]
        paidAt:
          type: string
    PaymentInput:
      type: object
      required: [amount, paymentMethod]
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
        notes:
          type: string
...